requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
cycler==0.11.0
fonttools==4.31.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.21.0
gym-notices==0.0.6
hydra-core==1.1.1
idna==3.3
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
my-gym @ git+https://github.com/osigaud/my_gym.git@5f08569a1c46425c74633d34c06c690ea32e7222
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pygame==2.1.2
pyparsing==3.0.7
pyre-extensions==0.0.23
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
-e git+https://github.com/facebookresearch/salina.git@ec9d114a5a0d79157430f59ac89a9171f5c67f35#egg=salina
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
xformers==0.0.10
